(maint) Add exec permissions issue to known issues
This adds a note about BOLT-1141, as we don't expect to prioritize it shortly and haven't heard of any users running into the issue. The issue is that shared files listed with the `files` key in task metadata don't retain their executable permissions when using the docker transport, which makes trying to call those executables fail.