(GH-2299) Write Windows local transport commands to file
This modifies the local transport to write commands to a file for
execution on Windows. This aligns with how we execute commands over the
WinRM transport, and enables two necessary behaviors: returning the exit
code to Bolt that was returned by the command (as opposed to simply 1 or
0, which is what our constructed command previously
returned) and to include pipes in the command which will be correctly
parsed in a file. This also ensures that Bolt commands behave the same
on the Windows local transport as over WinRM.

Closes #2299

!bug

* **Windows local transport returns correct exit codes and accepts pipes** ([#2299](https://github.com/puppetlabs/bolt/issues/2299))

  When running commands over the local transport on Windows machines,
  Bolt now returns the exit code returned by the command as opposed to
  just 0 or 1. It also accepts pipes as part of the command.